feat: EntityLoader 구현
